consolidate number formatting code in one place - associated cleanups - remove an old `dev` folder item pointing at missing files
